CTF BOT/requirements.txt
selenium
bs4
validate_email
discord.py
asyncio
lxml
py3DNS
pytz
